Windows VS2019 build for PostgreSQL 10.16 x86
The code as of sha:15e66cdeb1ca2feec28c02f54759dbbd340f98f0 was built
using Visual Studio 2019 and [winflexbison-2.5.24](https://github.com/lexxmark/winflexbison) against PostgreSQL 10.16-2 32-bit.

This build was only possible due to the work done by
[alexeyfadeev](https://github.com/alexeyfadeev/jsquery/tree/windows).
Thank you. Compared to that build, this makes use of the newest commit
to postgrespro/jsquery, while Alexey's build uses sources from 2017.

This build was not attempted using either newer versions of PostgreSQL
or using the 64-bit build for simplicities sake.

It was required to host a PostgreSQL database on Windows for use with [EHRbase](https://github.com/ehrbase/ehrbase), which requires the `jsquery` extensions and [`temporal_tables`](https://github.com/arkhipov/temporal_tables/releases/tag/v1.2.0), which is available for PostgreSQL 10.x as well (and can be downloaded as a pre-built DLL for Windows systems).